refactor: standardize entry points and project structure
- Move main logic to root index.js
- Update internal test imports to match new structure
- Update package.json entry points
