+++ package.json
@@ -2,7 +2,7 @@
   "name": "@fastify/session",
   "version": "11.1.1",
   "description": "a session plugin for fastify",
-  "main": "lib/fastifySession.js",
+  "main": "index.js",
   "type": "commonjs",
   "scripts": {
     "test": "npm run test:unit && npm run test:typescript",
